Move evals to separate directory
